requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file

django==3.2
djangorestframework==3.12
psycopg2==2.8.6
djangorestframework_simplejwt==4.6.0

